Added test fixes for move to video-web
