+++ VideoWeb/Testing.Common/Helpers/VideoWebUriFactory.cs
@@ -6,12 +6,23 @@ namespace Testing.Common.Helpers
     {
         public VideoWebParticipantsEndpoints ParticipantsEndpoints { get; }
         public VideoWebCallbackEndpoints CallbackEndpoints { get; }
+        public VideoWebConferenceEndpoints ConferenceEndpoints { get; }
 
         public VideoWebUriFactory()
         {
             ParticipantsEndpoints = new VideoWebParticipantsEndpoints();
             CallbackEndpoints = new VideoWebCallbackEndpoints();
+            ConferenceEndpoints = new VideoWebConferenceEndpoints();
         }
+    }
+
+    public class VideoWebConferenceEndpoints
+    {
+        private static string ApiRoot => "conferences";
+
+        public string GetConferencesForVhOfficer = $"{ApiRoot}/vhofficer";
+
+        public string GetConferenceById(Guid conferenceId) => $"/conferences/{conferenceId}";
     }
 
     public class VideoWebParticipantsEndpoints
