+++ VideoWeb/VideoWeb.AcceptanceTests/Builders/ExecuteRequestBuilder.cs
@@ -42,9 +42,15 @@ namespace VideoWeb.AcceptanceTests.Builders
             VerifyTheResponse();
         }
 
-        public void SendWithoutVerification()
+        public void SendToVideoApiWithoutVerification()
         {
             _context.Response = _context.VideoApiClient().Execute(_context.Request);
+            GetTheResponse();
+        }
+
+        public void SendToVideoWebWithoutVerification()
+        {
+            _context.Response = _context.VideoWebClient().Execute(_context.Request);
             GetTheResponse();
         }
