+++ VideoWeb/VideoWeb.AcceptanceTests/Configuration/ConferenceRetrievers/IConferenceRetriever.cs
@@ -1,0 +1,10 @@
+﻿using VideoWeb.AcceptanceTests.Contexts;
+using VideoWeb.Services.Video;
+
+namespace VideoWeb.AcceptanceTests.Configuration.ConferenceRetrievers
+{
+    public interface IConferenceRetriever
+    {
+        ConferenceDetailsResponse GetConference(TestContext context);
+    }
+}
