+++ VideoWeb/VideoWeb.AcceptanceTests/Strategies/HearingStatus/ClosedStrategy.cs
@@ -18,7 +18,7 @@ namespace VideoWeb.AcceptanceTests.Strategies.HearingStatus
             new ExecuteEventBuilder()
                 .WithContext(context)
                 .WithRequest(request)
-                .SendToVideoApi();
+                .SendToVideoWeb();
         }
     }
 }
