+++ VideoWeb/VideoWeb.AcceptanceTests/Strategies/HearingStatus/SuspendedStrategy.cs
@@ -18,7 +18,7 @@ namespace VideoWeb.AcceptanceTests.Strategies.HearingStatus
             new ExecuteEventBuilder()
                 .WithContext(context)
                 .WithRequest(request)
-                .SendToVideoApi();
+                .SendToVideoWeb();
         }
     }
 }
